requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
openai==0.28.1
propcache==0.4.1
py==1.11.0
python-dotenv==1.2.1
requests==2.32.5
retry==0.9.2
soupsieve==2.8
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
